Adding Parser.ExpectUntil to guard against infinite loops
