+++ .github/workflows/build_and_test.yml
@@ -21,4 +21,4 @@ jobs:
     - name: Build
       run: dotnet build src --no-restore
     - name: Test
-      run: dotnet test src --no-build --verbosity normal
+      run: dotnet test src --no-build --verbosity normal --blame-hang --blame-hang-timeout 30s
