+++ src/Todl.Compiler.Tests/CodeAnalysis/ParserTests/ParserTests.cs
@@ -6,8 +6,6 @@ using Xunit;
 
 namespace Todl.Compiler.Tests.CodeAnalysis
 {
-    using ArgumentsList = CommaSeparatedSyntaxList<Argument>;
-
     public sealed partial class ParserTests
     {
         private static TExpression ParseExpression<TExpression>(string sourceText)
@@ -77,30 +75,6 @@ namespace Todl.Compiler.Tests.CodeAnalysis
 
             var memberName = inputText[(inputText.LastIndexOf('.') + 1)..^0];
             memberAccessExpression.MemberIdentifierToken.Text.Should().Be(memberName);
-        }
-
-        [Fact]
-        public void TestParseBlockStatementBasic()
-        {
-            var inputText = @"
-            {
-                a = 1 + 2;
-                ++a;
-                b = a;
-            }
-            ";
-            var blockStatement = ParseStatement<BlockStatement>(inputText);
-
-            blockStatement.OpenBraceToken.Text.Should().Be("{");
-            blockStatement.OpenBraceToken.Kind.Should().Be(SyntaxKind.OpenBraceToken);
-
-            blockStatement.InnerStatements.Should().SatisfyRespectively(
-                _0 => _0.As<ExpressionStatement>().Should().NotBeNull(),
-                _1 => _1.As<ExpressionStatement>().Should().NotBeNull(),
-                _2 => _2.As<ExpressionStatement>().Should().NotBeNull());
-
-            blockStatement.CloseBraceToken.Text.Should().Be("}");
-            blockStatement.CloseBraceToken.Kind.Should().Be(SyntaxKind.CloseBraceToken);
         }
 
         [Fact]
